requirements.txt
pydantic>=2.0.0
replicate>=0.25.0
python-dotenv>=1.0.0
e2b-code-interpreter>=0.0.10

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
